backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sentence-transformers==2.2.2
pinecone-client==2.2.4
openpyxl==3.1.2

